==> scripts/09_make-live-stats.rmd <==
---
output:
  html_document:
    keep_md: TRUE
    theme: readable
    includes:
      before_body: 08_before-body-back-to-index-href.txt
---

```{r setup, include = FALSE, cache = FALSE}
knitr::opts_chunk$set(error = TRUE)
library(plyr)
library(yaml)
library(knitr)
suppressPackageStartupMessages(library(ggplot2))
```

```{r, echo = FALSE, message = FALSE}
mlu_teams <- read.delim(file.path("..", "data", "mlu-teams.tsv"),
                        stringsAsFactors = FALSE)
mlu_cols <- with(mlu_teams, setNames(color, team))
theme_set(theme_bw())

# when compiled via the control script, game identifier will have been
# read from command line and pre-stored in the environment
# for interactive use and development, set the game here
if(!exists("game")) {
  #game <- "2014-04-12_vanNH-at-pdxST"
  #game <- "2014-04-20_sfoDF-at-vanNH"
  #game <- "2014-05-10_seaRM-at-vanNH"
  #game <- "2014-05-31_vanNH-at-seaRM"
  #game <- "2014-06-07_seaRM-at-vanNH"
  #game <- "2014-06-15_pdxST-at-vanNH"
  #game <- "2014-04-12_seaRM-at-sfoDF"
  #game <- "2014-05-03_sfoDF-at-pdxST"
  #game <- "2014-04-26_vanNH-at-seaRM"
  #game <- "2014-06-15_pdxST-at-vanNH"
  #game <- "2014-06-14_pdxST-at-seaRM"
  #game <- "2014-04-19_bosWC-at-wdcCT"
  #game <- "2014-05-04_phlSP-at-wdcCT"
  #game <- "2014-06-21_bosWC-at-nykRM"
  #game <- "2014-07-19_vanNH-at-wdcCT"
  game <- "2015-05-23_seaRM-at-vanNH"
}

message(game, " > 09_vanNH-nowPlaying.rmd:")

## parse the game identifier
tmp <- strsplit(game, split = "_")[[1]]
game_date <- tmp[1]
tmp <- strsplit(tmp[2], split = "-")[[1]]
away_team <- tmp[1]
home_team <- tmp[3]
jTeams <- sort(c(away_team, home_team))

## define some groups of codes used in multiple places
goal_codes <- c('G', 'LG', 'CG')
assist_codes <- c('A', 'LA', 'PUA', 'DA')
d_codes <- c('D', 'HB', 'FB', 'DA')
keeper_codes <- c(goal_codes, assist_codes, d_codes)

## read in pre-computed player stats
pl_stats_file <- file.path("..", "games", game, "07_pass-game",
                           paste0(game, "_player-stats.tsv"))
pl_stats <- read.delim(pl_stats_file,
                       colClasses = list(player = "character", 
                                         last = "character",
                                         number = "character"))
pl_stats$pl_long <- with(pl_stats, paste(player, last, sep = "-"))
#str(pl_stats, max.level = 0)
pl_stats$team_pretty <-
  mapvalues(pl_stats$team, from = mlu_teams$team,
            to = mlu_teams$longName, warn_missing = FALSE)
#str(pl_stats)
pl_stats_by_team <- split(pl_stats, pl_stats$team)
vars_to_drop <- c('game', 'team', 'number', 'pl_long', 'team_pretty')
vars_to_show <- setdiff(names(pl_stats), vars_to_drop)
```

# `r away_team` at `r home_team` `r game_date`

```{r, echo = FALSE, message = FALSE}
build_roster_path <- function(team) {
  file.path("..", "rosters", paste0("2015_", team, "-roster.tsv"))}
away_roster_file <- build_roster_path(away_team)
home_roster_file <- build_roster_path(home_team)
read_roster <- function(roster_file) {
  x <- read.delim(roster_file, stringsAsFactors = FALSE,
                  ## do this because of seaRM-rupp-00
                  colClasses = list(number = "character"))
  x[c('number', 'last')]
}
away_roster <- read_roster(away_roster_file)
home_roster <- read_roster(home_roster_file)
message("  roster read for ", away_team, " and ", home_team)

yaml_file <- file.path("..", "games", game, paste0(game, "_game-info.yaml"))
game_yaml <- yaml.load_file(yaml_file)
message("  game YAML read; game status: ", game_yaml$game_status)

game_file <- file.path("..", "games", game, "06_possess-game",
                       paste0(game, "_gameplay-resolved.tsv"))
sDat <- read.delim(game_file,
                   colClasses = list(pl_team = "character",
                                     pl_pnum = "character",
                                     pl_code = "character"))
message("  ", nrow(sDat), " rows of game play found")

point_file <- file.path("..", "games", game, "06_possess-game",
                       paste0(game, "_points-resolved.tsv"))
point_info <- read.delim(point_file, stringsAsFactors = FALSE,
                         colClasses = list(clk_before = "character",
                                           clk_after = "character"))
message("  ", nrow(point_info), " rows of point info play found")

## set factor levels uniformly for all relevant 'team' variables
jFun <- function(x, jLevels = jTeams) factor(x, levels = jLevels)
sDat <- transform(sDat, pl_team = jFun(pl_team))
point_info <- transform(point_info, pull_team = jFun(pull_team),
                        scor_team = jFun(scor_team))

## save a version of sDat to display as raw data
sDat_safe <- sDat

## keep only D's, goals, and assists
sDat <- subset(sDat, pl_code %in% keeper_codes)
#str(sDat)
#table(sDat$pl_code)

## get back at least a dummy row for every point; necessary if a point has no
## game play with a keeper code
sDat <-
  suppressMessages(join(point_info[c('point', 'period', 'pull_team')], sDat))

## look up pl_pnum's last name
sDat$pl_lname <-
  I(with(sDat,
         ifelse(pl_team == away_team,
                away_roster$last[match(sDat$pl_pnum,
                                       away_roster$number)],
                home_roster$last[match(sDat$pl_pnum,
                                       home_roster$number)])))

## deal with cases where we have no last name
## how does this happen?
## * we use a number that's not on the roster
## * we don't capture the player's number and enter ? instead
## * the dummy row created for an EOP point
## this seems to be triggered by a plain ? ... check into this
name_unknown <- is.na(sDat$pl_lname) & !is.na(sDat$pl_team)
if(any(name_unknown)) {
  name_unknown <- which(name_unknown)
  sDat$pl_lname[name_unknown] <- "?name?"
  message("  ALERT last name unknown for player(s) involved in these plays:")
  message(with(sDat[name_unknown, ],
               paste("  point", point, "event", event, "team", pl_team, "number",
                     pl_pnum, "code", pl_code, collapse = "\n")))
  }

## fuse pl_team and pl_lname for humans to read
sDat$player <- with(sDat, factor(paste(pl_team, pl_pnum, pl_lname, sep = "-")))

## bring the score into sDat
sDat <- suppressMessages(join(sDat, point_info[c('point', jTeams)]))

## DONE WITH EVENT LEVEL DATA!

## prepare text reporting game status
whereInGame <-
  switch(game_yaml$game_status,
         complete = "game is complete",
         upcoming = "game is upcoming -- any data here is fake and for set up purposes",
         paste("period", sDat$period[nrow(sDat)]))

## move this to a script that checks player stats computed in
## 07_resolve-passes.r

## tally events by player
# player_dat <- ddply(sDat, ~ player + pl_team, function(z) {
#   goals <- sum(z$pl_code %in% goal_codes)
#   assists <- sum(z$pl_code %in% assist_codes)
#   Ds  <-  sum(z$pl_code %in% d_codes)
#   c(points = goals + assists,
#     goals = goals, assists = assists, Ds = Ds)
#   })
# player_dat_by_team <- split(player_dat, player_dat$pl_team)
# player_dat_by_team <- llply(player_dat_by_team, function(z) {
#   z <- arrange(z, desc(points), desc(goals), desc(assists), desc(Ds))
#   rownames(z) <- NULL
#   subset(z, select = -pl_team)
# })
# message("  player stat tables include ", nrow(player_dat_by_team[[away_team]]),
#         " for ", away_team, " and ", nrow(player_dat_by_team[[home_team]]),
#         " for ", home_team)
# 
# player_dat_by_team_combined <- do.call("rbind", player_dat_by_team)
# player_dat_by_team_combined <-
#   data.frame(game = game, player_dat_by_team_combined )
# 
# message("  writing player stats to file")
# 
# out_dir <- file.path("..", "games", game, "09_html")
# out_file <- file.path(out_dir, paste0(game, "_player-stats.tsv"))
# write.table(player_dat_by_team_combined, out_file,
#             quote = FALSE, sep = "\t", row.names = FALSE)

## prepare game scoring progression table
jFun <- function(z) {
  n <- nrow(z)
  pt <- z$point[1]
  period <- z$period[1]
  away_score <- z[[away_team]][n]
  home_score <- z[[home_team]][n]
  #comment <- z$comment[nrow(z)]
  
  pt_dur <- unlist(point_info[point_info$point == pt,
                              grep("pt_dur", names(point_info))])
  point_duration_text <- unname(pt_dur["pt_dur_txt"])
  
  time_before_after <-
    unname(unlist(point_info[point_info$point == pt,
                             grep("clk", names(point_info))]))
  
  is_a_goal <- z$pl_code %in% goal_codes
  if(any(is_a_goal)) {
    goal_row <- max(which(is_a_goal))
    scorer <- z$player[goal_row]
    is_an_assist <- z$pl_code %in% assist_codes
    if(sum(is_an_assist) < 1) { 
      assister <- NA ## must be a callahan, right?
      } else if(sum(is_an_assist) == 1) {
        assist_row <- max(which(is_an_assist))
        assister <- z$player[is_an_assist]
        } else {
          assister <- "???"
          message("point", point, ": more than 1 assist found?!?\n")
}                  
    desc <- paste(assister, "to", scorer)
    } else { ## time ran out with no goal
      desc <- "- no goal -"
      }
  ret_val <- c(period = period, 
               begin = time_before_after[1],
               end = time_before_after[2],
               pt_duration = point_duration_text,
               desc = desc,
               away_score = away_score,
               home_score = home_score)
  ret_val <-
    rename(ret_val, c("away_score" = away_team,
                      "home_score" = home_team))
  return(ret_val)
  }
gDat <- ddply(sDat, ~ point, jFun)
gDat <- gDat[rev(seq_len(nrow(gDat))), ]
message("  scoring progression table has ", nrow(gDat), " rows\n")
```

# `r away_team` `r sDat[[away_team]][nrow(sDat)]`
# `r home_team` `r sDat[[home_team]][nrow(sDat)]`
## `r whereInGame`

Go to ...  
  * [Scoring progression](#scoringProgression)  
  * [Player stats via figures](#pl_figs)  
  * [Player stats for `r away_team`](#away)  
  * [Player stats for `r home_team`](#home)  
  * [Data on goals, assists, D's](#selectData)  
  * [Full raw data](#rawData)  

## Scoring progression<a id="scoringProgression"></a>:
```{r echo = FALSE, results = 'asis'}
kable(gDat, row.names = FALSE)
```

## Player stats via figures<a id="pl_figs"></a>:

```{r function_to_prep_barchart, echo = FALSE, include = FALSE}
prep_for_barchart <- function(x, stat_var = "goals", abs_min = 1,
                              rank_max = 15) {
  x <- x[x[[stat_var]] >= abs_min, ]
  x$pl_long <- reorder(x$pl_long, x[[stat_var]])
  return(x)
  }
```

### Points = goals + assists

```{r barchart_points, echo = FALSE, warning = FALSE, fig.height = 8}
tmp <- ddply(pl_stats, ~ team, prep_for_barchart, stat_var = "points")
p <- ggplot(tmp, aes(x = pl_long, y = points, fill = team))
p + geom_bar(stat = "identity") + coord_flip() +
  #geom_text(aes(label = last), hjust = -0.15, size = 3.5) +
  geom_text(aes(label = points), hjust = -0.15, size = 3.5) +
  scale_fill_manual(values = mlu_cols) +
  ylim(0, max(tmp$points) * 1.05) +
  theme(legend.position = c(1, 0), legend.justification = c(0.88, 0.1),
        legend.background = element_rect(fill = 0),
        axis.text.x = element_text(size = rel(1)),
        axis.ticks = element_blank(), axis.title.y = element_blank())
```

### Goals

```{r barchart_goals, echo = FALSE, warning = FALSE, fig.height = 8}
tmp <- ddply(pl_stats, ~ team, prep_for_barchart, stat_var = "goals")
p <- ggplot(tmp, aes(x = pl_long, y = goals, fill = team))
p + geom_bar(stat = "identity") + coord_flip() +
  geom_text(aes(label = goals), hjust = -0.15, size = 3.5) +
  scale_fill_manual(values = mlu_cols) +
  ylim(0, max(tmp$goals) * 1.05) +
  theme(legend.position = c(1, 0), legend.justification = c(0.88, 0.1),
        legend.background = element_rect(fill = 0),
        axis.text.x = element_text(size = rel(1)),
        axis.ticks = element_blank(), axis.title.y = element_blank())
```

### Assists

```{r barchart_assists, echo = FALSE, warning = FALSE, fig.height = 8}
tmp <- ddply(pl_stats, ~ team, prep_for_barchart, stat_var = "assists")
p <- ggplot(tmp, aes(x = pl_long, y = assists, fill = team))
p + geom_bar(stat = "identity") + coord_flip() +
  geom_text(aes(label = assists), hjust = -0.15, size = 3.5) +
  scale_fill_manual(values = mlu_cols) +
  ylim(0, max(tmp$assists) * 1.05) +
  theme(legend.position = c(1, 0), legend.justification = c(0.88, 0.1),
        legend.background = element_rect(fill = 0),
        axis.text.x = element_text(size = rel(1)),
        axis.ticks = element_blank(), axis.title.y = element_blank())
```

### D's
```{r barchart_ds, echo = FALSE, warning = FALSE, fig.height = 8}
tmp <- ddply(pl_stats, ~ team, prep_for_barchart, stat_var = "def")
p <- ggplot(tmp, aes(x = pl_long, y = def, fill = team))
p + geom_bar(stat = "identity") + coord_flip() +
  geom_text(aes(label = def), hjust = -0.15, size = 3.5) +
  scale_fill_manual(values = mlu_cols) +
  ylim(0, max(tmp$def) * 1.05) +
  theme(legend.position = c(1, 0), legend.justification = c(0.88, 0.1),
        legend.background = element_rect(fill = 0),
        axis.text.x = element_text(size = rel(1)),
        axis.ticks = element_blank(), axis.title.y = element_blank())
```

## Player stats for `r away_team` <a id="away"></a>:

points = goals + assists  
tables sorted in decreasing order based on points (then goals, assists, Ds)  
Ds = total of plain ol' D's + interceptions (D), hand blocks (HB), and foot blocks (FB)  
stats are cumulative for this game

```{r results='asis', echo = FALSE, eval = FALSE}
if(nrow(player_dat_by_team[[away_team]]) > 0) {
  kable(player_dat_by_team[[away_team]], row.names = TRUE)
  }
```

```{r results='asis', echo=FALSE}
if(nrow(pl_stats_by_team[[away_team]]) > 0) {
  kable(pl_stats_by_team[[away_team]][vars_to_show],row.names = TRUE)
  }
```

## Player stats for `r home_team` <a id="home"></a>:

points = goals + assists  
tables sorted in decreasing order based on points (then goals, assists, Ds)  
Ds = total of plain ol' D's + interceptions (D), hand blocks (HB), and foot blocks (FB)  
stats are cumulative for this game

```{r results='asis', echo = FALSE, eval = FALSE}
if(nrow(player_dat_by_team[[home_team]]) > 0) {
  kable(player_dat_by_team[[home_team]], row.names = TRUE)
  }
```

```{r results='asis', echo=FALSE}
if(nrow(pl_stats_by_team[[home_team]]) > 0) {
  kable(pl_stats_by_team[[home_team]][vars_to_show], row.names = TRUE)
  }
```

## Raw data on select events (goals, assists, D's)<a id="selectData"></a>:

```{r results = 'asis', echo = FALSE}
kable(sDat, row.names = FALSE)
```

## Full raw data<a id="rawData"></a>:

__Note: just for display purposes.__ Raw data in more useful forms can be found in the [GitHub repository](https://github.com/jennybc/vanNH). Find the game you're interested in in the `games` subdirectory.

```{r results='asis', echo=FALSE}
kable(sDat_safe, row.names = FALSE)
```


```{r, echo=FALSE, results="hide", eval = FALSE}
## one off analysis
## looking for double happinesses
dhDat <- dlply(sDat, .(point),
               function(z) {
                 dGuys <- z$player[z$event == "D"]
                 scorer <- z$player[z$event == "G"]
                 if(length(scorer) > 0) {
                   if(scorer %in% dGuys) {
                     return(scorer)
                     }
                   }
                 }
               )
```
