==> scripts/25_player-stats-by-team.rmd <==
---
output:
  html_document:
    toc: yes
    keep_md: TRUE
---

<a href="../index.html">Back to index</a>

```{r include = FALSE, message = FALSE}
library(plyr)
library(knitr)
library(ggplot2)

theme_set(theme_bw())
mlu_teams <- read.delim(file.path("..", "data", "mlu-teams.tsv"),
                        stringsAsFactors = FALSE)
mlu_cols <- with(mlu_teams, setNames(color, team))

knitr::opts_chunk$set(error = TRUE)

pl_stats <- readRDS(file.path("..", "games",
                              "2014_all-games", "2014_player-game-stats.rds"))
str(pl_stats) # 1400 obs. of  14 variables:

pl_stats_agg <-
  readRDS(file.path("..", "games",
                    "2014_all-games", "2014_player-stats-aggregated.rds"))
str(pl_stats_agg) # 244 obs. of  13 variables:

# when compiled via the control script, team will have been
# read from command line and pre-stored in the environment
# for interactive use and development, set the team here
if(!exists("team")) {
  #team <- "wdcCT"
  team <- "vanNH"
  }

message(team, " > 25_player-stats.rmd:")

pl_stats <- pl_stats[pl_stats$team == team, ]
message("  ", nrow(pl_stats), " rows of player * game stats found")
#str(pl_stats, max.level = 1)

pl_stats_agg <- pl_stats_agg[pl_stats_agg$team == team, ]
message("  ", nrow(pl_stats_agg), " rows of game-aggregated player stats found")
str(pl_stats_agg) # 30 obs. of  13 variables

pl_stats$pl_long <- with(pl_stats, paste(player, last, sep = "-"))
pl_stats_agg$pl_long <- with(pl_stats_agg, paste(player, last, sep = "-"))
```

# Player statistics for `r team`

## Data

Tables and figures below based on data from `r length(unique(pl_stats$game))` games. The dataset covers `r length(unique(pl_stats$player))` unique players. There are `r nrow(pl_stats)` rows of player-level statistics, each being a unique combination of a game and a player. 

## Player stat table
```{r echo = FALSE, results = 'asis'}
vars_to_show <- c('last', 'player', 'games', 'points', 'goals', 'assists',
                  'throws', 'completions', 'comp_pct',
                  'def', 'catches', 'drop')
kable(pl_stats_agg[vars_to_show], row.names = FALSE)
```

## Scatterplot: completion percentage vs number of throws
```{r comp_pct_vs_throws, echo = FALSE, warning = FALSE}
set.seed(3)
p <- ggplot(subset(pl_stats_agg, !grepl("\\?", last)),
            aes(x = throws, y = comp_pct, label = last))
p <- p + geom_text(aes(colour = throws), size = 6, fontface = "bold",
                   position = position_jitter(width = 0, height = 0.007)) +
  guides(colour = FALSE) +
  scale_colour_gradient(low = "#D8D8D8", high = "#404040") +
  coord_cartesian(xlim = max(pl_stats_agg$throws) * c(-0.15, 1.15)) +
  xlab("total throws") + ylab("completion percentage") 
p
```

## Scatterplot: completion percentage vs throws per game
```{r comp_pct_vs_throws_per_game, echo = FALSE, warning = FALSE}
set.seed(3)
p <- ggplot(subset(pl_stats_agg, !grepl("\\?", last)),
            aes(x = throws/games, y = comp_pct, label = last))
p <- p + geom_text(aes(colour = throws/games), size = 6, fontface = "bold",
                   position = position_jitter(width = 0, height = 0.007)) +
  guides(colour = FALSE) +
  scale_colour_gradient(low = "#D8D8D8", high = "#404040") +
  coord_cartesian(xlim = with(pl_stats_agg, max(throws/games)) *
                    c(-0.15, 1.15)) +
  xlab("average throws per game") + ylab("completion percentage") 
p
```

```{r function_to_complete_share_of, echo = FALSE, include = FALSE}
prep_for_share_of_figure <- function(x, stat_var = "goals", abs_min = 1,
                                     rank_max = 15, label_share_min = 0.05) {
  tot_of_stat <- sum(x[[stat_var]])
  x <- x[x[[stat_var]] >= abs_min & rank(-x[[stat_var]]) <= rank_max, ]
  x$pl_long <- reorder(x$pl_long, x[[stat_var]])
  x$stat_share <- x[[stat_var]]/tot_of_stat
  x$share_pretty <- with(x, ifelse(stat_share > label_share_min,
                                   sprintf("%0.2f", stat_share), ''))
  x$last <- with(x, ifelse(stat_share > label_share_min, last, ''))
  share_var_name <- paste(stat_var, "share", sep = "_")
  x <- rename(x, c("stat_share" = share_var_name))
  return(x)
}
```

## Barchart: player's share of total season goals
```{r barchart_share_of_goals, echo = FALSE, warning = FALSE}
tmp <- prep_for_share_of_figure(pl_stats_agg, stat_var = "goals")
p <- ggplot(tmp, aes(x = pl_long, y = goals_share, fill = team))
p + geom_bar(stat = "identity") + coord_flip() +
  geom_text(aes(label = last), hjust = -0.15, size = 3.5) + 
  geom_text(aes(label = share_pretty), hjust = 1.15, size = 3.5) +
  scale_fill_manual(values = mlu_cols) +
  theme(axis.ticks = element_blank(), axis.title.y = element_blank()) +
  ylim(0, max(tmp$goals_share) * 1.15) +
  ylab("player's share of team goals")  + guides(fill = FALSE)
```

## Barchart: player's share of total season assists
```{r barchart_share_of_assists, echo = FALSE, warning = FALSE}
tmp <- prep_for_share_of_figure(pl_stats_agg, stat_var = "assists")
p <- ggplot(tmp, aes(x = pl_long, y = assists_share, fill = team))
p + geom_bar(stat = "identity") + coord_flip() +
  geom_text(aes(label = last), hjust = -0.15, size = 3.5) + 
  geom_text(aes(label = share_pretty), hjust = 1.15, size = 3.5) +
  scale_fill_manual(values = mlu_cols) +
  theme(axis.ticks = element_blank(), axis.title.y = element_blank()) +
  ylim(0, max(tmp$assists_share) * 1.15) +
  ylab("player's share of team assists")  + guides(fill = FALSE)
```

## Barchart: player's share of total season D's
```{r barchart_share_of_ds, echo = FALSE, warning = FALSE}
tmp <- prep_for_share_of_figure(pl_stats_agg, stat_var = "def")
p <- ggplot(tmp, aes(x = pl_long, y = def_share, fill = team))
p + geom_bar(stat = "identity") + coord_flip() +
  geom_text(aes(label = last), hjust = -0.15, size = 3.5) + 
  geom_text(aes(label = share_pretty), hjust = 1.15, size = 3.5) +
  scale_fill_manual(values = mlu_cols) +
  theme(axis.ticks = element_blank(), axis.title.y = element_blank()) +
  ylim(0, max(tmp$def_share) * 1.15) +
  ylab("player's share of team D's")  + guides(fill = FALSE)
```

## Scatterplot: player's share of total season goals vs D's
```{r share_of_goals_vs_ds, echo = FALSE, warning = FALSE}
tot_goals <- sum(pl_stats_agg$goals)
tot_ds <- sum(pl_stats_agg$def)
tmp <- subset(pl_stats_agg, def > 0 | goals > 0)
tmp <- mutate(tmp, def_share = def / tot_ds, goals_share = goals / tot_goals)
set.seed(3)
p <- ggplot(tmp, aes(x = def_share, y = goals_share, label = last))
p + geom_text(aes(colour = goals_share + 1.5 * def_share),
              size = 6, fontface = "bold",
              position = position_jitter(width = 0, height = 0.007)) + 
  guides(colour = FALSE) +
  scale_colour_gradient(low = "#D8D8D8", high = "#404040") +
  coord_cartesian(xlim = with(tmp, max(def_share)) * c(-0.15, 1.15)) +
  xlab("player's share of team D's") + ylab("player's share of team goals") +
  theme(strip.text = element_text(face = "bold", size = rel(1.25)))
```

```{r assist_perc_vs_incomp_perc, echo = FALSE, warning = FALSE, eval = FALSE}
## Parinella plot
tmp <- subset(pl_stats_agg, assists > 0,
              select = c(pl_long, throws, comp_pct, games, assists, last))
tmp <- mutate(tmp,
              thr_per_game = throws/games,
              incomp_pct = 1 - comp_pct,
              assist_share = assists/throws)
tmp
p <- ggplot(tmp, aes(x = incomp_pct, y = assist_share))
p + geom_point()
p + geom_text(aes(label = last))
## to do: if I finish this, use thr_per_game to designate low vs high touch
```
